policy-routing: initial import
This is inspired by the package Luci/freifunk-poilcyrouting [1] and uses
the same uci-file.
The Berlin community has found some issues, open questions about what's
going on and generally feels a bit unhappy with this scripts and their
logic.

So this is a proof of consept restart ...

this inital version is drafted in a node which is meshing only (not
own uplink to the internet). The meshing node will provide an uplink
via an non-standard-IpSec-tunnel.

[1] - https://github.com/openwrt/luci/tree/master/contrib/package/freifunk-policyrouting